requirements.txt
Flask==2.1.2
flask_cors==3.0.10
flask_marshmallow==0.14.0
flask_sqlalchemy==2.5.1
keras==2.9.0
numpy==1.23.0
python_bcrypt==0.3.2
tensorflow==2.9.1
Werkzeug==2.1.2
gunicorn
